[#78] Fix :merge 작업 수행
